requirements.txt
streamlit
pandas
numpy
openpyxl
pyarrow
pytest
oracledb
polars
python-dotenv